requirements/production.txt
Django==1.6.2
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0.2
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.19
bleach==1.4.1
celery==3.1.17
django-celery==3.1.16
django-debug-toolbar==1.2.2
django-extensions==1.4.9
django-grappelli==2.5.2
gnureadline==6.3.3
gunicorn==19.3.0
html5lib==0.999
ipython==2.3.1
kombu==3.0.24
pingpp==2.0.0
pytz==2014.10
raven==4.0.4
requests==2.5.1
six==1.9.0
sqlparse==0.1.14
upyun==2.2.5
wsgiref==0.1.2
xmltodict==0.9.1
